Reorganize arg bundles in the runtime.

(Reviewed and XC/ugni performance sanity-checked by @ronawho.)

The bundled arguments for outlined tasking and on-stmt body functions
consist of a runtime-defined header struct, followed by the arguments
themselves. There are distinct header struct types for tasking and
on-stmt arg bundles.

On-stmt, or "comm", bundle header structs have two uses. They support
comm layers using Active Messages to invoke on-stmt function bodies on
remote nodes, of course, but they also support internal Active Messages
for support activities such as coordinating program shutdown and doing
GETs and PUTs that cannot use comm layer RMA features. In comm=gasnet
the opcode, if you will, of an Active Message is delivered out-of-band
from the AM's contents. But in comm=ugni and comm=ofi the underlying
communication library either doesn't have such support or we don't use
it, so the opcode has to be embedded in the AM message itself. This
being so, it has to be in the same place in all AM messages, so that the
target-side AM handler can pick it out of the received message and
figure out how to proceed from there.

It used to be the case that the header struct for an on-stmt arg bundle
began with a member that was the header struct for a tasking arg bundle
and then followed that with some additional information. Thus all arg
bundles, of either kind, effectively began with the tasking header.
This allowed tasking layers to find their information in either kind of
bundle, which is necessary because the tasking layer actually runs the
body functions of both tasking and on-stmts. (All comm layers initiate
on-stmt body functions using the tasking layer on the target node.) The
downside of this was that comm arg bundles for internal AM opcodes, as
opposed to actual on-stmts, were excessively large because they had a
task bundle header struct on the front that they didn't need.

Here, reorganize the arg bundle headers. Put a small member encoding
the bundle header type first in both kinds of bundles, with room in the
encoding for a reasonable number of comm layer-specific internal AM
opcodes. This allows identifying the type of the bundle just by looking
at this initial member. Then, rearrange the contents of the comm header
so that the comm info comes first and the task header comes second. Now
task arg bundles still consist of a tasking header struct followed by
the bundled arguments, and on-stmt arg bundles consist of a comm header
struct followed by a tasking bundle (tasking header + bundled arguments).

This resolves Cray/chapel-private#926 and resolves #12350 and
resolves #12364.

+++ runtime/include/comm/chpl-comm-task-decls.h
@@ -24,18 +24,20 @@
 #ifndef _chpl_comm_task_decls_h
 #define _chpl_comm_task_decls_h
 
+#include <stdint.h>
+
 // Define the type of a n.b. communications handle.
 typedef void* chpl_comm_nb_handle_t;
 
 typedef struct {
-  int dummy;    // structs must be nonempty
+  int8_t dummy;    // structs must be nonempty
 } chpl_comm_taskPrvData_t;
 
 //
 // Comm layer private area within executeOn argument bundles
 // (bundle.comm)
 typedef struct {
-  int64_t dummy;    // structs must be nonempty
+  int8_t dummy;    // structs must be nonempty
 } chpl_comm_bundleData_t;
 
 #undef HAS_CHPL_CACHE_FNS
